Use 8bit sound samples for SDL
No need for 16bit audio with our crappy nearest upsampling.

+++ src/ui/sdl2/audio.rs
@@ -67,8 +67,8 @@ impl ::ui::Audio for Audio {
     }
 }
 
-// Use 16bit sound samples
-type Sample = i16;
+// Use 8bit sound samples
+type Sample = u8;
 
 /// Audio output sample rate
 const SAMPLE_RATE: u32 = 44_100;
